+++ BraintreeApi/src/main/java/com/braintreepayments/api/Braintree.java
@@ -462,14 +462,29 @@ public class Braintree {
 
     /**
      * Starts the Pay With PayPal flow. This will launch a new activity for the PayPal mobile SDK.
+     *
      * @param activity the {@link android.app.Activity} to receive the {@link android.app.Activity#onActivityResult(int, int, android.content.Intent)}
      *   when payWithPayPal finishes.
      * @param requestCode the request code associated with this start request. Will be returned
      * in {@code onActivityResult}.
      */
     public void startPayWithPayPal(Activity activity, int requestCode) {
+        startPayWithPayPal(activity, requestCode, null);
+    }
+
+    /**
+     * Starts the Pay With PayPal flow. This will launch a new activity for the PayPal mobile SDK.
+     *
+     * @param activity the {@link android.app.Activity} to receive the {@link android.app.Activity#onActivityResult(int, int, android.content.Intent)}
+     *   when payWithPayPal finishes.
+     * @param requestCode the request code associated with this start request. Will be returned
+     * @param additionalScopes A {@link java.util.List} of additional scopes. Ex: PayPalOAuthScopes.PAYPAL_SCOPE_ADDRESS.
+     * Acceptable scopes are defined in {@link com.paypal.android.sdk.payments.PayPalOAuthScopes}.
+     * in {@code onActivityResult}.
+     */
+    public void startPayWithPayPal(Activity activity, int requestCode, List<String> additionalScopes) {
         sendAnalyticsEvent("add-paypal.start");
-        mBraintreeApi.startPayWithPayPal(activity, requestCode);
+        mBraintreeApi.startPayWithPayPal(activity, requestCode, additionalScopes);
     }
 
     /**